Fix: motivate browsers to reload web app resources
by mangling the Git commit hash into the MD5 hash of all web application
resource's ETag headers, we ensure that index.html is essentially
invalidated in the browser's cache, even if the actual content is the
same between firmware versions. this hopefully makes the browser
recognize that it has to reload app.js in particular.